Put shared components in the separate directory under shared
