+++ src/app/pages/services/opcua/opcua.component.ts
@@ -5,8 +5,8 @@ import { LocalDataSource } from 'ng2-smart-table';
 
 import { OpcuaService } from 'app/common/services/opcua/opcua.service';
 import { NotificationsService } from 'app/common/services/notifications/notifications.service';
-import { ConfirmationComponent } from 'app/shared/confirmation/confirmation.component';
-import { DetailsComponent } from 'app/shared/details/details.component';
+import { ConfirmationComponent } from 'app/shared/components/confirmation/confirmation.component';
+import { DetailsComponent } from 'app/shared/components/details/details.component';
 import { MessagesService } from 'app/common/services/messages/messages.service';
 import { OpcuaStore } from 'app/common/store/opcua.store';
 import { FsService } from 'app/common/services/fs/fs.service';
