+++ src/app/pages/user-groups/user-groups.component.ts
@@ -7,8 +7,8 @@ import { UserGroup } from 'app/common/interfaces/mainflux.interface';
 import { UserGroupsService } from 'app/common/services/users/groups.service';
 import { FsService } from 'app/common/services/fs/fs.service';
 import { NotificationsService } from 'app/common/services/notifications/notifications.service';
-import { ConfirmationComponent } from 'app/shared/confirmation/confirmation.component';
-import { DetailsComponent } from 'app/shared/details/details.component';
+import { ConfirmationComponent } from 'app/shared/components/confirmation/confirmation.component';
+import { DetailsComponent } from 'app/shared/components/details/details.component';
 
 const defFreq: number = 100;
