+++ src/app/shared/components/chart/chart.colors.ts
