+++ src/app/shared/components/confirmation/confirmation.component.scss
