+++ src/app/shared/components/message-table/message-table.component.scss
